requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Collection
requests==2.31.0
urllib3==2.1.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2

# Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
python-dateutil==2.8.2

# Optional (for advanced features)
# scipy==1.11.4
# scikit-learn==1.3.2